% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildDiseaseNetwork}
\alias{buildDiseaseNetwork}
\title{Build disease-specific PPI network}
\usage{
buildDiseaseNetwork(network, nodeWeight, nodeAnnot, nodeNameBy,
  nodeSelected = NULL)
}
\arguments{
\item{network}{A PPI network (Pajek format graph)}

\item{nodeWeight}{Node weight (per-gene association p value) and Ensemble gene symbol data frame}

\item{nodeAnnot}{Node annotation (data frame)}

\item{nodeNameBy}{Gene Symbol option. It takes either "Ensemble" or "HUGO".}

\item{nodeSelected}{By default, we use the entire network nodes (nodeSelected = NULL). The set of nodes can be limited to a subset of the entire network.}
}
\value{
A disease-specific PPI network (Pajek format graph)
}
\description{
Provides a phenotype-specific PPI network: The nodes in the PPI network are assigned weights reflect#'ing the association with a phenotype. The weights are given by -log(p) + log(p_sigThres).
}

