% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calEdgePot}
\alias{calEdgePot}
\title{calculate the edge potential}
\usage{
calEdgePot(sigSubnet, sigSubnet.adj, nodeStates, All.cells)
}
\arguments{
\item{sigSubnet}{the top subnetwork graph}

\item{sigSubnet.adj}{the adjcency matrix of the top subnetwork}

\item{nodeStates}{the tissue specificity matrix for the top subnetwork genes}

\item{All.cells}{all the cells in the tissue specificity matrix}
}
\value{
the edge potentials of a UGM (data frame)
}
\description{
calculate the edge potential of UGM
}

