% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calEdgeStates}
\alias{calEdgeStates}
\title{calculate edge states}
\usage{
calEdgeStates(g2, g2.adj, edgePot, nodeStates, cell)
}
\arguments{
\item{g2}{the top subnetwork}

\item{g2.adj}{the adjacency matrix of top subnetwork}

\item{edgePot}{the edge potentails of the top subnetwork edges}

\item{nodeStates}{on/off states of the top subnetwork nodes}

\item{cell}{a cell/tissue}
}
\value{
edge states a vector of all the edge potentials
}
\description{
calculate edge states
}

