% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calJointPDF2}
\alias{calJointPDF2}
\title{calculate joint PDFs of a UGM model}
\usage{
calJointPDF2(diseaseNetwork, topSubnet, edgePotList, nodeStatesList, logZvals,
  nNull)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{topSubnet}{top subnetworks}

\item{edgePotList}{the edge potentials for the top subnetworks}

\item{nodeStatesList}{the on/off states of nodes fro the top subnetworks}

\item{logZvals}{the logarithmic value of the partition function}

\item{nNull}{the number of null top subnetworks}
}
\value{
the joint PDFs of the top subnetworks
}
\description{
calculate joint probability distribution (PDF) of a UGM model
}

