% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calKStest}
\alias{calKStest}
\title{two sample KS test}
\usage{
calKStest(x, y, option, nboots = NULL, alternative)
}
\arguments{
\item{x}{A test set}

\item{y}{A reference set}

\item{option}{either "standard" or "bootstrap"}

\item{nboots}{number of bootstrap}

\item{alternative}{"less" or "greater"}
}
\value{
KS test p values
}
\description{
perform two sample KS test either the statdard KS test or bootstrap version
}

