% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calMeanInverseSPD}
\alias{calMeanInverseSPD}
\title{calculate mean inverse shortest path distance}
\usage{
calMeanInverseSPD(sigSubnet, nodeStates, cell)
}
\arguments{
\item{sigSubnet}{the top subnetwork}

\item{nodeStates}{the on/off states of nodes}

\item{cell}{a cell/tissue}
}
\value{
mean inverse shortest path distance
}
\description{
calculate mean Inverse Shortest path distance among activeNodes
}

