% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calPartitionfunction}
\alias{calPartitionfunction}
\title{calculate partiontion function}
\usage{
calPartitionfunction(nNull, resultDir, UGMPATH, MATLABPATH)
}
\arguments{
\item{nNull}{the number of times to calculate the partition function}

\item{resultDir}{a dirctory where the result is saved}

\item{UGMPATH}{a path for the UGM software}

\item{MATLABPATH}{a path for MATLAB software}
}
\value{
a set of calculated partition function
}
\description{
calculate partition function of a UGM (Undirected Graphical Model)
}

