% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calSPD}
\alias{calSPD}
\title{get Shortest path distance among a set of nodes}
\usage{
calSPD(graph, nodelist)
}
\arguments{
\item{graph}{a network/graph}

\item{nodelist}{a set of nodes}
}
\value{
shortest path distance

all the nodes in shortest path distance
}
\description{
get Shortest path distance among a set of nodes
}

