% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coexpressAnalysis}
\alias{coexpressAnalysis}
\title{coexpression Analysis}
\usage{
coexpressAnalysis(coexDB, coexAnnot, nodeAnnotPath, topSubnets, nNull)
}
\arguments{
\item{coexDB}{Coexpression DB}

\item{coexAnnot}{Entrez ID annoation}

\item{nodeAnnotPath}{Node annotation file}

\item{topSubnets}{Top sub network genes}

\item{nNull}{The number of null top subnetworks}
}
\value{
a coexpression coefficient matrix
}
\description{
coexpression Analysis
}

