% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convIWIDtoEntrez}
\alias{convIWIDtoEntrez}
\title{convert InWeb ID to Entrez ID}
\usage{
convIWIDtoEntrez(IWID, nodeAnnotPath, coexAnnot, coexDB)
}
\arguments{
\item{IWID}{InWeb ID}

\item{nodeAnnotPath}{Node annotation file}

\item{coexAnnot}{Entrez ID annoation}

\item{coexDB}{Coexpression DB}
}
\value{
InWeb ID to Entrez ID conversion table
}
\description{
convert InWeb ID to Entrez ID
}

