% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evalNodeScore}
\alias{evalNodeScore}
\title{Evaluete node scores}
\usage{
evalNodeScore(pvalues, sigThres, nIterate = 0)
}
\arguments{
\item{pvalues}{Per-gene association score for a given phenotype}

\item{sigThres}{Genome-wide association threshold. e.g., sigThres <- 5e-6}

\item{nIterate}{It allows iterative disease subnetwork search. Use the default value "nIterate <- 0" for the top subnetwork search.'Please keep the default value until the option is completed.}
}
\value{
calibrated association scores
}
\description{
Evaluete node scores
}

