% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evalSigIntersect}
\alias{evalSigIntersect}
\title{evaluate intersect significance}
\usage{
evalSigIntersect(set1TotalSize, set2TotalSize, intersectSize, TotalPopulation)
}
\arguments{
\item{set1TotalSize}{The size of the first set}

\item{set2TotalSize}{The size of the second set}

\item{intersectSize}{The sized of intersect between the two sets}

\item{TotalPopulation}{The size of total population where the two sets are sampled}
}
\value{
p value of hypergeometic test
}
\description{
evaluate the significance of the intersect of the two sampled sets from a total population.
}

