% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCoexCoeff}
\alias{getCoexCoeff}
\title{get coexpression coefficient}
\usage{
getCoexCoeff(id1, id2, coexDB)
}
\arguments{
\item{id1}{a Entrez ID}

\item{id2}{a Entrez ID}

\item{coexDB}{Coexpression DB}
}
\value{
a coexpression coefficient
}
\description{
get coexpression coefficient
}

