% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCoexCoeffMat}
\alias{getCoexCoeffMat}
\title{get coexpression coefficient matrix}
\usage{
getCoexCoeffMat(entrezID, coexDB)
}
\arguments{
\item{entrezID}{a set of Entrez ID}

\item{coexDB}{Coexpression DB}
}
\value{
a coexpression coefficient matrix
}
\description{
get coexpression coefficient matrix
}

