% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadGeneExpression}
\alias{loadGeneExpression}
\title{load gene expression data}
\usage{
loadGeneExpression(expressionDataFile, nodeAnnot, nodeNameby = "Ensemble")
}
\arguments{
\item{expressionDataFile}{gene expression data file (tissue specificity)}

\item{nodeAnnot}{Node annotation (data frame)}

\item{nodeNameby}{Gene Symbol option. It takes either "Ensemble" or "HUGO"}
}
\value{
gene expression data frame
}
\description{
Load gene expression data
}

