% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadNodeWeight}
\alias{loadNodeWeight}
\title{Load node weights (association p values)}
\usage{
loadNodeWeight(nodeWeight, nodeAnnot, nodeNameby)
}
\arguments{
\item{nodeWeight}{Node weight file}

\item{nodeAnnot}{Node annotation (data frame)}

\item{nodeNameby}{Gene Symbol option. It takes either "Ensemble" or "HUGO".}
}
\value{
nodeWeigtTable (a data frame sorted according to InWeb node annotation.)
}
\description{
Load node weights (e.g. per gene association p values) for a protein protein interaction network.
The input file is required to have "Ensemble" or "HUGO" field as a gene symbol and "pvalues" field as per-gene association.
We provide per gene scores from mutationally constraint study as an example. It is located under /Data.
}

