% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadPPINetwork}
\alias{loadPPINetwork}
\title{Load protein protein interaction network}
\usage{
loadPPINetwork(networkfile, nodeAnnotfile)
}
\arguments{
\item{networkfile}{Pajek format graph file}

\item{nodeAnnotfile}{A file that contains the network node symbols and annotations. i.e, Ensemble/HUGO gene symbol}
}
\value{
a graph object with node names annotated by "Ensemble" gene symbol

a data frame with node annotation (Ensemble and HUGO gene symbol etc.)
}
\description{
Load protein protein interaction network and annotate each node by Ensemble gene symbol (default).
We provide InWeb3 protein interaction network (Pajek graph format) and the annotation file (text table).
They are located under /Data. User can provide their own graph and the annotation file.
}

