% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mergeWithOrder}
\alias{mergeWithOrder}
\title{merge data frames while preserving the row order}
\usage{
mergeWithOrder(x, y, ..., sort = T, keep_order)
}
\arguments{
\item{x}{a data frame}

\item{y}{a data frame}

\item{...}{additional parameters to pass}

\item{sort}{sort option (T or F)}

\item{keep_order}{keep the row order based on either first or second data frame}
}
\value{
A merged data frame
export
}
\description{
merge data frames while preserving the row order
}
\references{
http://www.r-statistics.com/2012/01/merging-two-data-frame-objects-while-preserving-the-rows-order/
}

