% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotTissueActivity}
\alias{plotTissueActivity}
\title{plot the tissue activity}
\usage{
plotTissueActivity(diseaseNetwork, geneExpression, targetCells, topSubnets)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{geneExpression}{a data frame for the tissue specificity}

\item{targetCells}{A set of cells/tissues under test}

\item{topSubnets}{top subnetwork genes}
}
\value{
A network plot showing the tissue specific genes in a top subnetwork for target tissues
}
\description{
plot the tissue activity
}

