% Generated by roxygen2 (4.0.2): do not edit by hand
\name{randDiseaseNetwork}
\alias{randDiseaseNetwork}
\title{randomize disease-specific PPI network}
\usage{
randDiseaseNetwork(diseaseNetwork, nPermute,
  permuteOptions = "Non-Degree-preserving")
}
\arguments{
\item{diseaseNetwork}{A a phenotype-specific PPI network}

\item{nPermute}{number of permutations}

\item{permuteOptions}{There are two options. "Non-Degree Preserving" - Node degree is not constrained in randomization of nodes in PPI. "Degree Preserving" - Node degree is preserved in randomization of nodes in PPI.}
}
\value{
randomized disease network node assignment table (data frame)
}
\description{
Randomize disease-specific PPI network
}

