% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readUGMInput}
\alias{readUGMInput}
\title{read UGM inputs}
\usage{
readUGMInput(UGMresultDir, nNull)
}
\arguments{
\item{UGMresultDir}{UGM result directory}

\item{nNull}{number of times to calcualte the partition function}
}
\value{
a list of UGM constructs
}
\description{
read UGM inputs
}

