% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runDiseaseSubnetSearch}
\alias{runDiseaseSubnetSearch}
\title{run disease-associated subnetwork search}
\usage{
runDiseaseSubnetSearch(diseaseNetwork, nodeAnnot, sigThres, nIterate = 0,
  nullDiseaseNetwork = NULL, showResult = TRUE)
}
\arguments{
\item{diseaseNetwork}{A phenotype-specific PPI network}

\item{nodeAnnot}{Node annotation (data frame)}

\item{sigThres}{Genome-wide association threshold. e.g., sigThres <- 5e-6}

\item{nIterate}{It allows iterative disease subnetwork search. Use the default value "nIterate <- 0" for the top subnetwork search.}

\item{nullDiseaseNetwork}{null association disease networks (randomized disease-association network)}

\item{showResult}{show the top subnetwork search result}
}
\value{
top/null scoring subnetworks SubnetList[[1]] corresponds to top scoring subnet. SubnetList[[2:]] the null top scoring subnets.
}
\description{
run disease-associated subnetwork search
}

