% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showSubNetSearchResult}
\alias{showSubNetSearchResult}
\title{show subNet search result on the fly x}
\usage{
showSubNetSearchResult(diseaseNetwork, nodeAnnot, module, scores)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{nodeAnnot}{node annotation (data frame)}

\item{module}{the detected subnetwork}

\item{scores}{the node scores}
}
\value{
summary
}
\description{
show subNet search result on the fly
}

