% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simGeneExpression}
\alias{simGeneExpression}
\title{simulated gene expression data}
\usage{
simGeneExpression(nodeAnnot, allCells, topSubnet, targetCells,
  geneNameBy = "Ensemble")
}
\arguments{
\item{nodeAnnot}{node annotation data frame}

\item{allCells}{a set of cells/tissues}

\item{topSubnet}{the top subnetwork}

\item{targetCells}{a set of target cells/tissues}

\item{geneNameBy}{gene symbol options, e.g., "Ensemble" or "HUGO"}
}
\value{
simulated gene expression
}
\description{
simulate gene expression data
}

