% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simRandNodeWeight}
\alias{simRandNodeWeight}
\title{Simulate signal/null nodes weights}
\usage{
simRandNodeWeight(nNodes, nSigNodes, sigFun, nullFun, sigThres, ...)
}
\arguments{
\item{nNodes}{the number of PPI network nodes to simulate}

\item{nSigNodes}{the number of signal nodes}

\item{sigFun}{the signal distribution}

\item{nullFun}{the null distribution}

\item{sigThres}{the signal threshold}

\item{...}{additional parameters to pass}
}
\value{
a numeric vector of p values
}
\description{
Simulate per-gene node weights (p values) from signal + null distribution for PPI network nodes.
For example, the signal distribution can be sampled from a beta distribution and
the null distribution from a uniform distribution.
}

