% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simRandSubnet}
\alias{simRandSubnet}
\title{Simulate random subnetwork from PPI network}
\usage{
simRandSubnet(nNodes, nSigNodes, nodeAnnot, subnetMatrix, nSubnet, nPermute)
}
\arguments{
\item{nNodes}{the number of PPI network nodes}

\item{nSigNodes}{the number of signal nodes}

\item{nodeAnnot}{Node annotation (data frame)}

\item{subnetMatrix}{A sampled signal subnetwork. ex. /Data/randomSubnet_ex1.txt}

\item{nSubnet}{the number of sub network : Please use a default value for now, ie.,  nSubnet <- 1}

\item{nPermute}{number of permutations with the Subnet : Please use a default value for now, ie.,  nPermute <- 1}
}
\value{
node_table - A vector (matrix) of sampled nodes.

node_table[1:nSigNodes] - signal node index of the network

node_table[(nSigNodes+1):nNodes] - null node index of the network
}
\description{
Provides a random node assignment with a signal subnetwork. The signal subnetwork refers to a
connected subnetwork where all the nodes are assumed to be associated with a given phenotype in con#' sideration.
}

