% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testDiesClassEnrichment}
\alias{testDiesClassEnrichment}
\title{test disease class enrichment of gene sets}
\usage{
testDiesClassEnrichment(geneDiseaseclass, topSubnet)
}
\arguments{
\item{geneDiseaseclass}{a data frame for gene-Disease class matrix}

\item{topSubnet}{top subnetwork genes}
}
\value{
p values of Fisher's exact test and contingency table
}
\description{
test disease class enrichment of gene sets
}

