% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testOMIMEnrichment}
\alias{testOMIMEnrichment}
\title{test OMIM record enrichment of gene set}
\usage{
testOMIMEnrichment(geneOMIM, topSubnet)
}
\arguments{
\item{geneOMIM}{a data frame for gene-OMIM record}

\item{topSubnet}{top subnetwork genes}
}
\value{
p values of Fisher's exact test and contingency table
}
\description{
test OMIM record enrichment of gene set
}

