% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testRankEnrichment}
\alias{testRankEnrichment}
\title{test rank enrichment of gene sets}
\usage{
testRankEnrichment(diseaseNetwork, geneExpression, targetCells, topSubnet,
  TSSigThres)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{geneExpression}{a data frame for the tissue specificity}

\item{targetCells}{A set of cells/tissues under test}

\item{topSubnet}{top subnetwork genes}

\item{TSSigThres}{the significance cutoff for the tissue specificity}
}
\value{
a list of p values: KS standard test and KS bootstrap test
}
\description{
test rank enrichment of gene sets
}

