% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testSigClustering}
\alias{testSigClustering}
\title{Test the significance of association signal clustering}
\usage{
testSigClustering(diseaseNetwork, nodeAnnot, detectedSubnets, nullNetwork,
  nNull, sigThres)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{nodeAnnot}{Node annotation (data frame)}

\item{detectedSubnets}{A list of detected nodes for the top subnetworks}

\item{nullNetwork}{null association disease networks (randomized disease-association network)}

\item{nNull}{the number of null networks}

\item{sigThres}{genome-wide association threshold. e.g., sigThres <- 5e-6}
}
\value{
p values of the permutation test for node #, edge #, clustering coefficient, and genetic score
}
\description{
Test the significance of association signal clustering (top subnetwork) against null association signal clustering (null top subnetworks)test the significance of association signal clustering
}

