% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testSigProportion}
\alias{testSigProportion}
\title{test the significance of the proportion}
\usage{
testSigProportion(diseaseNetwork, geneExpression, targetCells, topSubnet,
  TSSigThres)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{geneExpression}{a data frame for the tissue specificity}

\item{targetCells}{A set of cells/tissues under test}

\item{topSubnet}{top subnetwork genes}

\item{TSSigThres}{the significance cutoff for the tissue specificity}
}
\value{
p values of Fisher's exact test
}
\description{
test the significance of the proportion
}

