% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testSigTissueCoherence}
\alias{testSigTissueCoherence}
\title{test the Significance of coordination of gene set (e.g., tissue specific genes)}
\usage{
testSigTissueCoherence(diseaseNetwork, geneExpression, targetCells, topSubnets,
  nNull)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{geneExpression}{a data frame for tissue specificity.}

\item{targetCells}{A set of cells/tissues under test.}

\item{topSubnets}{Top sub network genes}

\item{nNull}{number of null gene sets (randomly selected genes in a give tissue)}
}
\value{
p values of MISPD (mean inverse shortest path distance)
}
\description{
test the significance of the connectivity of a set of expressed genes for target tissues.
}

