% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testSigTissueOverlap}
\alias{testSigTissueOverlap}
\title{test the significance of the overlap among target cell/tissues.}
\usage{
testSigTissueOverlap(diseaseNetwork, geneExpression, targetCells, topSubnet)
}
\arguments{
\item{diseaseNetwork}{A phenotype-specific PPI network (Pajek format graph)}

\item{geneExpression}{A data frame for the tissue specificity.}

\item{targetCells}{A set of cells/tissues under test.}

\item{topSubnet}{Top subnetwork genes}
}
\value{
p values
}
\description{
test the significance of the overlap between gene sets from target cell/tissues.
}

