% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testSigTissueSpecificity}
\alias{testSigTissueSpecificity}
\title{Tissue specificity analysis}
\usage{
testSigTissueSpecificity(diseaseNetwork, geneExpression, topSubnets, nNull,
  permuteOption, algorithmOption = NULL, MATLABPATH = NULL)
}
\arguments{
\item{diseaseNetwork}{a phenotype-specific PPI network (Pajek format graph)}

\item{geneExpression}{a data frame tissue specificity.}

\item{topSubnets}{a list of the detected subnetworks}

\item{nNull}{the number of null top subnetworks}

\item{permuteOption}{"nullTopsubnet" or "randGeneExpression"}

\item{algorithmOption}{UGMPath (UGM software installation directory)}

\item{MATLABPATH}{MATLABPATH (Matlab installation directory)}
}
\value{
p values
}
\description{
test the significance of tissue specificity
}
\references{
UGM software
}

