% Generated by roxygen2 (4.0.2): do not edit by hand
\name{UGMConstruct1}
\alias{UGMConstruct1}
\title{UGM construct with null top subnetwork}
\usage{
UGMConstruct1(diseaseNetwork, geneExpression, topSubnets, nNull)
}
\arguments{
\item{diseaseNetwork}{A phenotype-specific PPI network}

\item{geneExpression}{a data frame for the tissue specificity}

\item{topSubnets}{a list of detected top subnetworks}

\item{nNull}{the number of null top subnetworks}
}
\value{
a list of UGM construct outputs
}
\description{
UGM construct with null top subnetwork
}

