% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeUGMResult}
\alias{writeUGMResult}
\title{writeUGMResult}
\usage{
writeUGMResult(UGM, resultDir)
}
\arguments{
\item{UGM}{A list of UGM constructs}

\item{resultDir}{A directory where the UGM constructs are written to}
}
\description{
writeUGMResult
}

